refactor(core): move builtin ops to their own file (#10336)
