+++ core/lib.rs
@@ -9,6 +9,7 @@ mod module_specifier;
 mod modules;
 mod normalize_path;
 mod ops;
+mod ops_builtin;
 mod ops_json;
 pub mod plugin_api;
 mod resources;
@@ -52,8 +53,6 @@ pub use crate::modules::ModuleSourceFuture;
 pub use crate::modules::NoopModuleLoader;
 pub use crate::modules::RecursiveModuleLoad;
 pub use crate::normalize_path::normalize_path;
-pub use crate::ops::op_close;
-pub use crate::ops::op_resources;
 pub use crate::ops::serialize_op_result;
 pub use crate::ops::Op;
 pub use crate::ops::OpAsyncFuture;
@@ -64,6 +63,8 @@ pub use crate::ops::OpResponse;
 pub use crate::ops::OpState;
 pub use crate::ops::OpTable;
 pub use crate::ops::PromiseId;
+pub use crate::ops_builtin::op_close;
+pub use crate::ops_builtin::op_resources;
 pub use crate::ops_json::op_async;
 pub use crate::ops_json::op_sync;
 pub use crate::resources::Resource;
